deployments/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Jung Dong Ho <dongho971220@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.1"
PyJWT = "^2.1.0"
cryptography = "^3.4.7"
requests = "^2.25.1"
requests-oauthlib = "^1.3.0"
jupyterlab = "^3.0.16"
psycopg2-binary = "^2.8.6"
SQLAlchemy = "^1.4.17"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
